backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
edge-tts
httpx
aiofiles
ffmpeg-python
python-multipart
requests
pillow
wheel
